real kind=8

+++ examples/test_derivauto.f90
@@ -11,11 +11,11 @@ PROGRAM test_derivauto
 
 !  For vars routine (called below)
 !  "vars" Output variables:
-   REAL(kind=r8), DIMENSION(1) :: ph, pco2, fco2, co2, hco3, co3, OmegaA, OmegaC, BetaD, rhoSW, p, tempis
+   REAL(kind=8), DIMENSION(1) :: ph, pco2, fco2, co2, hco3, co3, OmegaA, OmegaC, BetaD, rhoSW, p, tempis
 !  "vars" Input variables
    INTEGER :: N
-   REAL(kind=r8), DIMENSION(1) :: temp, sal, alk, dic, sil, phos, Patm, depth, lat
-   REAL(kind=r8), DIMENSION(6,1) :: ph_deriv, pco2_deriv, fco2_deriv, co2_deriv, &
+   REAL(kind=8), DIMENSION(1) :: temp, sal, alk, dic, sil, phos, Patm, depth, lat
+   REAL(kind=8), DIMENSION(6,1) :: ph_deriv, pco2_deriv, fco2_deriv, co2_deriv, &
                      hco3_deriv, co3_deriv, omegaa_deriv, omegac_deriv
 !  "vars" Input options
    CHARACTER(10) :: optCON, optT, optP, optB, optKf, optK1K2
@@ -23,8 +23,8 @@ PROGRAM test_derivauto
 
 !  Local variables:
    INTEGER :: i
-   REAL(kind=r8) :: H
-   REAL(kind=r8), DIMENSION(6) :: H_deriv
+   REAL(kind=8) :: H
+   REAL(kind=8), DIMENSION(6) :: H_deriv
    CHARACTER*4 :: invar(6)
    
    invar(1) = 'Alk '
