+++ examples/test_kprime.f90
@@ -10,10 +10,10 @@ PROGRAM test_kprime
    INTEGER, PARAMETER :: n = 1
    
 !  Output variables:
-   REAL(kind=r8), DIMENSION(1) :: kp_cfc11, kp_cfc12, kp_sf6
+   REAL(kind=8), DIMENSION(1) :: kp_cfc11, kp_cfc12, kp_sf6
    
 !  Input variables
-   REAL(kind=r8), DIMENSION(1) :: temp, sal
+   REAL(kind=8), DIMENSION(1) :: temp, sal
    
 !  Input at standard T and S
 !  temp(1)   = 25.0
