+++ examples/test_kzero.f90
@@ -10,10 +10,10 @@ PROGRAM test_kzero
    INTEGER, PARAMETER :: n = 1
    
 !  Computed variables:
-   REAL(kind=r8), DIMENSION(1) :: k0_co2, k0_n2o
+   REAL(kind=8), DIMENSION(1) :: k0_co2, k0_n2o
    
 !  Input variables
-   REAL(kind=r8), DIMENSION(1) :: temp, sal
+   REAL(kind=8), DIMENSION(1) :: temp, sal
    
 !  Input at standard T and S
 !  temp(1)   = 25.0
