+++ src/p80.f90
@@ -25,17 +25,17 @@ FUNCTION p80(dpth,xlat)
 
 ! Input variables:
   !> depth [m]
-  REAL(kind=r8), INTENT(in) :: dpth
+  REAL(kind=8), INTENT(in) :: dpth
   !> latitude [degrees]
-  REAL(kind=r8), INTENT(in) :: xlat
+  REAL(kind=8), INTENT(in) :: xlat
 
 ! Output variable:
   !> pressure [db]
-  REAL(kind=r8) :: p80
+  REAL(kind=8) :: p80
 
 ! Local variables:
-  REAL(kind=r8) :: pi
-  REAL(kind=r8) :: plat, d, c1
+  REAL(kind=8) :: pi
+  REAL(kind=8) :: plat, d, c1
 
   pi=3.141592654
